Update README.md
In this project, I will be using a cyber security SIEM tool called Splunk. Splunk is a powerful software tool that is used to monitor and troubleshoot a variety of systems. I will be doing a variety of tasks using Splunk such as; uploading a file or data into Splunk, performing a basic search, evaluating the fields, narrowing my findings, searching for a failed login for root, and evaluating the search result.